+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Change Log
+## [0.0.3] - 2021-04-06
+- Added List view mode 🤩
+- Improved responsiveness when updating sidebar size
+- Improved color contrast
+
 ## [0.0.2] - 2021-03-23
 - Updated README.md